requirements.txt
requests>=2.32.3
bs4>=0.0.2
beautifulsoup4>=4.13.3
Flask>=3.1.0
waitress>=3.0.2
sentry-sdk[flask]>=2.22.0
python-dotenv>=1.0.1
yt-dlp>=2025.2.19